-prepare rutil for merge to main  - windows build fixes  - added missing license blocks
git-svn-id: https://svn.resiprocate.org/rep/resiprocate/branches/b-TKLC-perf_work@9334 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ rutil/AbstractFifo.hxx
@@ -439,13 +439,13 @@ class AbstractFifo : public FifoStatsInterface
 
             if(mCounter >= 4096)
             {
-               mAverageServiceTimeMicroSec=resipIntDiv(diff, mCounter);
+               mAverageServiceTimeMicroSec=(UInt32)resipIntDiv(diff, mCounter);
             }
             else // fifo got emptied; merge into a rolling average
             {
                // .bwc. This is a moving average with period 64, round to 
                // nearest int.
-               mAverageServiceTimeMicroSec=resipIntDiv(
+               mAverageServiceTimeMicroSec=(UInt32)resipIntDiv(
                      diff+((4096-mCounter)*mAverageServiceTimeMicroSec),
                      4096U);
             }
